refactor: update query to get needed columns; update mapper with columns

+++ src/main/java/com/sweaterbank/leasing/car/model/UserLease.java
@@ -8,7 +8,7 @@ public record UserLease (
         ApplicationStatus status,
         Date applicationDate,
         String interestBase,
-        float leaseMargin,
+        BigDecimal leaseMargin,
         BigDecimal monthlyPayment,
         BigDecimal borrowedAmount
 ) { }
